Remove use of username
Remove use of username. The username isn't used but does complicate the
process, as it enforces knowledge of the type of user Object that's
given. As the typing is not enforced by Spring, this would complicate
the dispatch interceptor significantly. Furthermore, the username isn't
used on the handling end and as such it only acts as additional
information instead of enforcing the roles.

#3530