Handle redirects manually to maintain headers
As a security measure, the HttpURLConnection does not keep headers
during a redirect. This causes a multitude of issues with the Update
Checker, most obviously being the los of module information (due to
which the checker cannot perform its intended task). Hence, we should
manually check for a redirect status code and if encountered, should
construct a new connection and set the headers again. To not end in a
never ending redirect loop, we max the redirect count to 10 (for now).

#3565