Introduce a ConversionCache and use it in the GenericMessage
Introduce a ConversionCache and use it in the GenericMessage. The
ConversionCache acts as a simplification for the GenericMessage, as an
additional check is required before returning a previous conversion.
Namely, we need to ensure that the given Converter was used for a Type
the payload was already converted in. If we don't do that, there's a
very real change that the conversion result comes out differently.

#3102