[#3098] feat(eventhandling): new EventProcessing configuration + `SequencingEventHandlingComponent` (#3573)
Main features introduced in this PR:
- `SequencingEventHandlingComponent` - that keeps sequencing handling by
sequence identifier, do not allow more than one event from the same
sequence to be handled concurrently
- changes in `EventProcessors` batch processing - now it doesn't depends
on the `UnitOfWork`, but just `ProcessingContext` - the UnitOfWork is
handled on the processor level itself.
- `EventProcessingConfigurer` - allows to define defaults configuration
for all event processors as well as just certain types (pooledStreaming,
subscribing), allows to register processor modules.

Removed:
- for now there is no tracing for batch processing level, but we need to
rewrite the whole tracing, because it's based on ThreadLocals. Maybe
tracing on UnitOfWork level will be enough.

TODO in the following PRs:
- move decorators of `EventHandlingComponent` out of the EventProcessor
modules.
- move all `SubscribingEventProcessor` related classes to the
`subscribing` package.