Fine tune interceptor support
Fine tune interceptor support. To that end, include an
HandlerInterceptorRegistry in the configuration, to which the
addInterceptor can add interceptors. The interceptor retrieval method
should provide the Configuration to extract the interceptors. This
return method should take the local interceptors *and* global
interceptors. Furthermore, move the intercepting event handling
component decoration to a separate decorator, for clarity. Lastly, add a
 test case to validate all this.

#3103