Adjust delegate transaction construction
Adjust delegate transaction construction. The InterceptingEventStore
does a computeResourceIfAbsent operation on the delegate#transaction
method. However, this operation typically also does a
computeResourceIfAbsent. Due to this, we land in timing related
concurrency exceptions on the concurrent map in the ProcessingContext.
To solve this, we do not use computeResourceIfAbsent in the
InterceptingEventStore, as described in the private method's javadoc

#3579