Introduce QueryGateway#queryMany
Introduce QueryGateway#queryMany. This operation acts as a replacement
of the QueryGateway#query operation allowing the ResponseType to be
provided. This shift will (1) ensure users don't have to deal with the
ResponseType anymore for direct queries, (2) make invocations on the
gateway clear, and (3) make the mapping in the DefaultQueryGateway a lot
 more straightforward.

#3488