[#3432] Make the `TokenStore` async-native (#3715)
This PR changes the API of the `TokenStore` to be async-native. 
We are therefore taking advantage of the major release, even though the
underlying implementation will continue to work synchronously for the
time being.
The API change essentially opens up the possibility of introducing both
imperative and reactive implementations at a later date.
As with other infrastructure components, a `ProcessingContext` can be
passed to each method.

This PR resolves #3432.

---------

Co-authored-by: Steven van Beelen <steven.vanbeelen@axoniq.io>
Co-authored-by: Mateusz Nowak <mateusz@nakodach.pl>