Replace AtomicBoolean for AtomicReference of final Configuration
Replace AtomicBoolean for AtomicReference of final Configuration. For
some somewhat unclear reason, we hit the
DefaultComponentRegistry#doBuild operation twice due to the
SpringLazyCreatingModule creating a Module. By simply returning the
previous Configuration (which is safe to do, as this operation is
invoked as part of a synchronized block) instead of throwing an
Exception.