Move legacy UnitOfWork and all the belongs with it to axon-legacy
Move legacy UnitOfWork and all the belongs with it to axon-legacy. We
need to keep it around for a bit longer due to the legacy-aggregate and
legacy-saga modules. Furthermore, clases in axon-todo also still expect
it existence. Adding it to axon-legacy ensures everything can reach it.
While moving the legacy UnitOfWork, some leftover classes that remained
in the to-be-released modules have been moved to legacy/todo modules as
well. Furthermore, its usage in the GenericMessage for correlation data
has been dropped, as required.

#3064