[#3828] fix(axon-server-connector): handle error query result
Explanation: There is some issue in the errors on the framework side, but that's a different issue. It starts with the fact that we should not use "completeWithError" for this case. We must just do a sendLast with a response that carries an ErrorMessage.
When we "completeWithError", AxonServer sees that as a failure for the handler to handle the query at all, and will try do divert the query to another handler. If that doesn't happen, it will reply with "no handler found".