Removed `SubscriptionQueryMessage` and responseType from QueryMessage
This change simplifies the query handling structure by eliminating the `SubscriptionQueryMessage` interface, its implementations, and associated tests. Query handling now relies on the `QueryMessage` interface, streamlining the codebase.

It also removes the requirement for Queries to specify the type of response they expect. This should all be encoded in the contract for a particular query via its fully qualified name

Solves #3831