Fix potential concurrency issue during split and merge
When splitting and merging, the Coordinator keeps hold of the claim on a token, but removes the Worker from scheduling. On the next coordination task, it may happen that the coordinator recognizes a claimed segment and a missing worker, and creates a new worker for it. This may cause events to be handled twice.

This commit adds a forced release to avoid workers being created for the segments that are being split or merged.