Remove the command handling result as a return value
Remove the command handling result as a return value. This one line
causes the entire event handler to wait for the command handling to
succeed. We should add either (1) a wait reset in the
AccessSerializingRepository to unblock or (2) introduce another
unblocking mechanism in the registered Event Handler (guessing the
SimpleEventHandlingComponent or the WorkPackage)

#3929