Remove unused module `spring-boot-autoconfiguration`
Probably caused by a merge of 4.12.x we had an unused module, not part of the build, containing one single class: `UpdateCheckConfiguration`.

The class was renamed to `UpdateCheckerAutoConfiguration` and moved to the spring boot extension. Checking for changes, I saw the `disabled`property was added, but as it is already present in the extension auto-configuration, I believe it is safe just to remove the unsed module.