Resolve message types for query handlers
The AnnotatedQueryHandlingComponent did not take into account the message-annotations (such as `Query`) when determining the message name for the registration. This causes a mismatch (No handler found for query), as the QueryGateway does this upon sending.

This PR introduces the MessageTypeResolver for the AnnotatedQueryHandlingComponent and uses it to determine the message type.