Introduce ArchUnit into commons module (#4015)
Adds an ArchUnit test to commons to ensure we're not introducing package
cycles or package hierarchy violations into this modules. This is the
first of such tests to such detect problems earlier in the development
process.

- Package cycles indicate that a package is not focused enough on a
specific purpose. Package cycles make it harder when you want move
packages to a new or a different module. They are often the first
indication of high level design mistakes, or that a newly introduced or
modified class is taking on too many responsibilities.
- Package hierarchy violations are when more specialized packages rely
on more general packages. Usually the more general package is an API
while the specialized package is an implementation. To make it easier to
later be able to publish there seperate, the API should live in a
separate branch in the package structure.

So instead of:

- com.foo/Database.java (API)
- com.foo/pg/PgDatabase.java (refers to a com.foo)
- com.foo/oracle/OracleDatabase.java (refers to a com.foo)

Keep the API in a separate branch:

- com.foo.api/Database.java (API)
- com.foo.pg/PgDatabase.java (refers to a com.foo.api)
- com.foo.oracle/OracleDatabase.java (refers to a com.foo.api)

This now enables you to split off the API module more easily, without
having to include all its implementations.

This PR freezes the current state, so no new cycles or hierarchy
violations can be introduced.