#3784: extend inspection for polymorphic entities with entity member on concrete types
We now also check the concrete types of an event sourced entity for having member entities, which requires a different inspection of the type in the AnnotatedEntityMetamodel.

Without this check, having a record as concrete type in an event sourced entity (either a sealed interface implementation or a concrete type declared in `@EventSourcedEntity`) having an `@EntityMember` field will result in duplicated handler registration (see org.axonframework.common.ReflectionUtils.deduplicateRecordFields).

Added also an additional Integration test to cover the case of a sealed interface with a concrete record as implementation.